requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4       
colorama==0.4.4
decorator==4.4.2
ffmpeg==1.4
fpstimer==0.0.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
numpy==1.21.1
opencv-contrib-python==4.5.3.56 
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==8.3.1
playsound==1.3.0
proglog==0.1.9
pygame==2.0.1
requests==2.26.0
tqdm==4.62.0
urllib3==1.26.6